feat: add GitHub Action, PyPI packaging, and GitLab MR comments
Restructure Iara from a monolithic script into a proper Python package
for distribution via pip and GitHub Marketplace. Key changes:

- Extract ai-codereview.py into iara/ package (config, prompt, reviewer,
  scanner, cli, models, extensions)
- Add pyproject.toml for PyPI distribution as 'iara-reviewer'
- Add GitHub Action (action.yml, Dockerfile, entrypoint.sh) with
  automatic PR commenting via GitHub API
- Update GitLab CI template to post review comments on MRs via API
- Keep ai-codereview.py as backwards-compatible shim
- Simplify test imports to use package directly
- Update README with GitHub, GitLab, and generic CI integration docs
