+++ .gitignore
@@ -1,4 +1,8 @@
 .env
 __pycache__/
 venv/
+*.egg-info/
+dist/
+build/
+.eggs/
