+++ iara/__main__.py
@@ -1,0 +1,5 @@
+"""Permite executar iara como modulo: python -m iara"""
+from iara.cli import main
+
+if __name__ == "__main__":
+    main()
