+++ pyproject.toml
@@ -1,0 +1,36 @@
+[build-system]
+requires = ["setuptools>=64"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "iara-reviewer"
+version = "1.0.0"
+description = "AI-powered code reviewer using OpenRouter LLMs"
+readme = "README.md"
+license = {text = "MIT"}
+requires-python = ">=3.8"
+authors = [{name = "Felipe Fernandes"}]
+keywords = ["code-review", "ai", "llm", "openrouter", "github", "gitlab"]
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Developers",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3",
+    "Programming Language :: Python :: 3.8",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Topic :: Software Development :: Quality Assurance",
+]
+
+[project.urls]
+Repository = "https://github.com/felipefernandes/iara"
+Issues = "https://github.com/felipefernandes/iara/issues"
+
+[project.scripts]
+iara = "iara.cli:main"
+
+[tool.setuptools.packages.find]
+include = ["iara*"]
